+++ CHANGELOG.md
@@ -21,12 +21,15 @@ New features:
 * Added possibility to resume timeouted jobs.
 * Added possibility to input hashes / upload hashlists with usernames.
 * Added possibility to input hashes / upload hashlists in "LM:NT" format (output format of tools like impacket).
+* Added column 'Username' in 'Job Info' -> 'Hashes'.
+* Added column 'Username' in 'Hashes'.
+* Added dropdown menu to specify input format.
 
 General improvements:
 * Improved/reimplemeted generator and scheduling logic. Better precision, more predictable workunit sizes. Default desired seconds per workunit changed to 600 seconds.
 * Improved benchmarking logic.
 * Improved 'Add New Job' UI, small redesign.
-* Added dropdown menu to specify input format.
+* Added support for cracking hashes with usernames.
 * It is now possible to upload dictionaries from any folders on the server.
 * It is now possible to upgrade standard (non-dockerized) Fitcrack to newer versions. Please see UPGRADE.md file.
 * Added new settings per hosts.